fastapi-template/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx==0.25.2
transformers==4.35.2
torch==2.1.1
numpy==1.24.3
opencv-python==4.8.1.78
Pillow==10.1.0
filetype==1.2.0
scikit-image==0.21.0
google-generativeai==0.3.2
python-multipart==0.0.6
pydantic==2.5.0
faiss-cpu==1.7.4
sentence-transformers==2.2.2
pdfplumber==0.10.3
PyMuPDF==1.23.8
google-cloud-vision==3.4.5
playwright==1.40.0
reportlab==4.0.7
